refactor: merge scenes and components
